# Try esbuild-loader to reduce Webpack compilation time