--- a/browser/package.json
+++ b/browser/package.json
@@ -50,6 +50,7 @@
   },
   "devDependencies": {
     "cypress": "^10.4.0",
+    "esbuild-loader": "^2.21.0",
     "jest-styled-components": "^6.0.0",
     "ts-migrate": "^0.1.30",
     "typescript": "^4.7.4"