Adding this Webpack loader ([esbuild-loader](https://github.com/privatenumber/esbuild-loader)) significantly reduces Webpack compilation time. For me, running `yarn start` in the `browser` directory now takes 15s to compile rather than 1 min.