@rileyhgrant 2023-02-03
Holy moly this is super great.

I experienced similarly huge reductions in compilation time to Matt. On my machine:
- `yarn start` went from ~45s -> ~12s
- `yarn build` went from ~60s -> ~30s

Crazy that a single plugin can give 4x faster compilation! Very nice find.